Fix linux c++ test build
Not sure why, but for some reason tests under `release/test` on Linux
fail to link if columnwise_summary.cpp is included in
libvisualization.a. Including it in the parent directory (unity_core)
seems to fix the issue; including the seemingly added dependencies of
libvisualization (like cppipc) does not, for some reason.