WIP: refactor columnwise summary to do all columns, smaller batch size
This will align it with the logic that all other transformers use, such that
a single batch contains a snapshot of the full state of the transformer.

WIP because it's not compiling yet - need to update all the other transformers
to treat batch_size as a runtime parameter instead of compile time.